+++ heron/instance/src/python/basics/spout_instance.py
@@ -22,7 +22,7 @@ from heron.common.src.python.utils.tuple import TupleHelper
 from heron.common.src.python.utils.metrics import SpoutMetrics
 from heron.common.src.python.utils.misc import SerializerHelper
 from heron.proto import topology_pb2, tuple_pb2
-from heron.streamparse.src.python import Stream
+from heron.pyheron.src.python import Stream
 
 import heron.common.src.python.constants as constants
