+++ heron/pyheron/src/python/component/component_spec.py
