+++ heron/pyheron/src/python/spout/base_spout.py
