+++ heron/pyheron/tests/python/stream_unittest.py
@@ -16,7 +16,7 @@
 # pylint: disable=protected-access
 import unittest
 
-from heron.streamparse.src.python.stream import Stream, Grouping
+from heron.pyheron.src.python.stream import Stream, Grouping
 from heron.proto import topology_pb2
 
 class StreamTest(unittest.TestCase):
